Backport premium test issue and minor bug from premium sources
